% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtd_combined_ROC_PR_plot.R
\name{combined_ROC_PR_plot}
\alias{combined_ROC_PR_plot}
\title{combined_ROC_PR_plot method}
\usage{
combined_ROC_PR_plot(self)
}
\arguments{
\item{self}{self}
}
\description{
combined_ROC_PR_plot method
}
