% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtd_hist_target_plotter.R
\name{hist_target_plotter}
\alias{hist_target_plotter}
\title{hist_target_plotter method}
\usage{
hist_target_plotter(self, target_name)
}
\arguments{
\item{self}{self}
}
\description{
hist_target_plotter method
}
