% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF1.R
\name{rf1}
\alias{rf1}
\title{Title: Brief description of rf1}
\usage{
rf1(
  bspid,
  prep,
  trees = 1000,
  grid = 11,
  n_boots = NA,
  seed = NA,
  min_max = 1/100,
  tune_metric = "roc_auc"
)
}
\arguments{
\item{bspid}{Description of parameter bspid.}

\item{prep}{Description of parameter prep}

\item{trees}{Description of parameter trees}

\item{grid}{Description of parameter grid}

\item{n_boots}{Description of parameter n_boots}

\item{seed}{Description of parameter seed}

\item{min_max}{Description of parameter min_max}
}
\value{
What the function returns.
}
\description{
A longer explanation of what rf1 does.
}
