% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF2.R
\name{rf2}
\alias{rf2}
\title{Title: Random Forest 2}
\usage{
rf2(
  bspid,
  prep,
  target,
  n = NA,
  trees = 1000,
  grid = 11,
  n_boots = 30,
  seed = NA,
  tune_metric = "pr_auc"
)
}
\arguments{
\item{bspid}{A character BSP-ID of Interest}

\item{prep}{A data.frame dataset}

\item{target}{Name of targe variable}

\item{n}{Optional: Number of rows to use}

\item{trees}{Optional: Number of trees}

\item{grid}{Optional: tuning grid}

\item{n_boots}{Optional: Number of resamples}

\item{seed}{Optional: seed}

\item{tune_metric}{Optional: Tuning metric "roc_auc", "pr_auc", "accuracy" etc.}
}
\value{
Object of class RF
}
\description{
Simplified resampling, case weighting and general efficiency.
}
