% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF3.R
\name{rf3}
\alias{rf3}
\title{Title: Random Forest 3}
\usage{
rf3(
  bspid,
  prep,
  target,
  n = NA,
  trees = 1000,
  grid = 11,
  n_boots = 30,
  seed = NA,
  tune_metric = "f_meas"
)
}
\arguments{
\item{bspid}{A character BSP-ID of Interest}

\item{prep}{A data.frame dataset}

\item{target}{Name of targe variable}

\item{n}{Optional: Number of rows to use}

\item{trees}{Optional: Number of trees}

\item{grid}{Optional: tuning grid}

\item{n_boots}{Optional: Number of resamples}

\item{seed}{Optional: seed}

\item{tune_metric}{Optional: Tuning metric "roc_auc", "pr_auc", "accuracy" "f_meas_weighted", etc.}
}
\value{
Object of class RF
}
\description{
Simplified resampling, case weighting and general efficiency.
}
