% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF5.R
\name{rf5}
\alias{rf5}
\title{Title: Random Forest 4}
\usage{
rf5(
  identity,
  data,
  id,
  id_vars,
  target_vars,
  target,
  row_limit = NA,
  trees = 1000,
  grid = 11,
  n_boots = 30,
  seed = NA,
  tune_metric = "mae",
  train_ratio = 0.8,
  n = 100,
  randomize_target = FALSE,
  testrun = FALSE,
  minimize_memory = FALSE
)
}
\arguments{
\item{data}{A data.frame dataset}

\item{id}{Name of id variable}

\item{target}{Name of target variable}

\item{row_limit}{Optional: Limit number of rows in data}

\item{trees}{Optional: Number of trees}

\item{grid}{Optional: Tuning grid}

\item{n_boots}{Optional: Number of resamples}

\item{seed}{Optional: Seed}

\item{tune_metric}{Optional: One metric of "mae", "kap"}

\item{train_ratio}{Optional: Portion of train data. Default: 0.80}

\item{n}{Optional: Lower limit of entries per target class}
}
\value{
Object of class RF
}
\description{
Simplified resampling, case weighting and general efficiency.
}
