% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF6_binary.R
\name{rf6_binary}
\alias{rf6_binary}
\title{Title: Random Forest 6 - Classification}
\usage{
rf6_binary(
  identity = list(),
  data,
  id,
  id_vars,
  target_vars,
  target,
  time_ax,
  row_limit = NA,
  trees = 1000,
  grid = 11,
  n_boots = 30,
  seed = NA,
  tune_metric = "pr_auc",
  train_ratio = 0.8,
  n = 30,
  randomize_target = FALSE,
  testrun = FALSE,
  quantiles = NA,
  min_sensitivity = 0.95
)
}
\arguments{
\item{target}{Name of targe variable}

\item{trees}{Optional: Number of trees}

\item{grid}{Optional: tuning grid}

\item{n_boots}{Optional: Number of resamples}

\item{seed}{Optional: seed}

\item{tune_metric}{Optional: Tuning metric "roc_auc", "pr_auc", "accuracy" "f_meas_weighted", etc.}

\item{n}{Optional: Number of rows to use}

\item{min_sensitivity}{Optional: Minimum sensitivity for minority class threshold selection (default 0.95)}

\item{bspid}{A character BSP-ID of Interest}

\item{prep}{A data.frame dataset}
}
\value{
Object of class RF
}
\description{
Simplified resampling, case weighting and general efficiency.
}
