% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RF6_multiclass.R
\name{rf6_multiclass}
\alias{rf6_multiclass}
\title{Title: Random Forest 6 - Multiclass Classification}
\usage{
rf6_multiclass(
  identity = list(),
  data,
  id,
  id_vars,
  target_vars,
  target,
  time_ax,
  row_limit = NA,
  trees = 1000,
  grid = 11,
  n_boots = 30,
  seed = NA,
  tune_metric = "pr_auc",
  train_ratio = 0.8,
  n = 30,
  randomize_target = FALSE,
  testrun = FALSE
)
}
\arguments{
\item{identity}{Optional: List with identity information}

\item{data}{A data.frame dataset}

\item{id}{Name of id variable}

\item{id_vars}{Names of id/grouping variables to exclude from predictors}

\item{target_vars}{Names of other target variables to exclude from predictors}

\item{target}{Name of target variable}

\item{time_ax}{Name of time axis variable}

\item{row_limit}{Optional: Limit number of rows in data}

\item{trees}{Optional: Number of trees}

\item{grid}{Optional: Tuning grid size}

\item{n_boots}{Optional: Number of resamples/folds}

\item{seed}{Optional: Seed}

\item{tune_metric}{Optional: Tuning metric "roc_auc", "accuracy", "f_meas", "pr_auc"}

\item{train_ratio}{Optional: Portion of train data. Default: 0.80}

\item{n}{Optional: Lower limit of entries per target class}

\item{randomize_target}{Optional: Randomize target variable (for baseline testing)}

\item{testrun}{Optional: Speed up for testing}
}
\value{
Object of class RF
}
\description{
Multiclass classification with group folds, case weighting and probability output.
}
